Title: 컴퓨터와 함께하는 숫자 게임 2, Time: 142ms, Memory: 10MB, Status: Passed - CodeTree

+++ 240822/컴퓨터와 함께하는 숫자 게임 2/play-number-game-with-computer-2.java
@@ -24,7 +24,7 @@ public class Main {
 
 	}
 
-	private static int binarySearch(long m, long target) {
+	private static long binarySearch(long m, long target) {
 
 		long turn = 1;
 		long l = 1, r = m;
